Fix compiler errors on macOS

+++ controller_interface/src/controller_interface_base.cpp
@@ -34,7 +34,7 @@ return_type ControllerInterfaceBase::init(
 
   try
   {
-    auto_declare<int>("update_rate", cm_update_rate);
+    auto_declare<int>("update_rate", static_cast<int>(cm_update_rate));
     auto_declare<bool>("is_async", false);
   }
   catch (const std::exception & e)
